레벨업 시스템,  UI작업 구조 완성

+++ Source/ProjectM/Character/Monster/MMonsterTypes.h
@@ -44,7 +44,7 @@ struct FMMonsterInfo
 	int32 MonsterReward = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
-	float ExperiencePoint = 0.f;
+	int64 ExperiencePoint = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
 	float MonsterMaxSpeed = 300.f;
